refactor: LottoTest와 LotteryTest의 분리 테스트 대상을 명확하게 함
